Separate execution types from messaging types
Plans, Protocols, Operations, States, and so forth aren't messaging
primitives, since they can be created and used locally. They *can* be
sent in messages, but don't necessarily need to be. Drawing this
distinction lays some of the groundwork for sorting out next generation
Plans and Protocols.
