+++ syft/__init__.py
@@ -51,10 +51,10 @@ from syft.frameworks.torch.fl import FederatedDataset, FederatedDataLoader, Base
 from syft.federated.train_config import TrainConfig
 
 # Import messaging objects
-from syft.messaging.protocol import Protocol
-from syft.messaging.plan import Plan
-from syft.messaging.plan import func2plan
-from syft.messaging.plan import method2plan
+from syft.execution.protocol import Protocol
+from syft.execution.plan import Plan
+from syft.execution.plan import func2plan
+from syft.execution.plan import method2plan
 
 # Import Worker Types
 from syft.workers.virtual import VirtualWorker
