+++ syft/execution/__init__.py
@@ -1,0 +1,3 @@
+from syft.execution.plan import func2plan
+from syft.execution.plan import method2plan
+from syft.execution.plan import Plan
