+++ syft/messaging/message.py
@@ -12,9 +12,9 @@ from syft.workers.abstract import AbstractWorker
 
 from syft.frameworks.torch.tensors.interpreters.placeholder import PlaceHolder
 
+from syft_proto.execution.v1.operation_pb2 import Operation as OperationPB
 from syft_proto.messaging.v1.message_pb2 import ObjectMessage as ObjectMessagePB
 from syft_proto.messaging.v1.message_pb2 import OperationMessage as OperationMessagePB
-from syft_proto.types.syft.v1.operation_pb2 import Operation as OperationPB
 from syft_proto.types.syft.v1.arg_pb2 import Arg as ArgPB
