+++ syft/serde/protobuf/proto.py
@@ -13,12 +13,15 @@ from syft.frameworks.torch.tensors.interpreters.placeholder import PlaceHolder
 from syft.generic.pointers.pointer_tensor import PointerTensor
 from syft.messaging.message import ObjectMessage
 from syft.messaging.message import Operation
-from syft.messaging.plan.plan import Plan
-from syft.messaging.protocol import Protocol
-from syft.messaging.plan.state import State
+from syft.execution.plan import Plan
+from syft.execution.protocol import Protocol
+from syft.execution.state import State
 
 
 from google.protobuf.empty_pb2 import Empty
+from syft_proto.execution.v1.plan_pb2 import Plan as PlanPB
+from syft_proto.execution.v1.protocol_pb2 import Protocol as ProtocolPB
+from syft_proto.execution.v1.state_pb2 import State as StatePB
 from syft_proto.frameworks.torch.tensors.interpreters.v1.additive_shared_pb2 import (
     AdditiveSharingTensor as AdditiveSharingTensorPB,
 )
@@ -28,9 +31,6 @@ from syft_proto.frameworks.torch.tensors.interpreters.v1.placeholder_pb2 import 
 from syft_proto.generic.pointers.v1.pointer_tensor_pb2 import PointerTensor as PointerTensorPB
 from syft_proto.messaging.v1.message_pb2 import ObjectMessage as ObjectMessagePB
 from syft_proto.messaging.v1.message_pb2 import OperationMessage as OperationMessagePB
-from syft_proto.messaging.v1.plan_pb2 import Plan as PlanPB
-from syft_proto.messaging.v1.protocol_pb2 import Protocol as ProtocolPB
-from syft_proto.messaging.v1.state_pb2 import State as StatePB
 from syft_proto.types.syft.v1.id_pb2 import Id as IdPB
 from syft_proto.types.torch.v1.device_pb2 import Device as DevicePB
 from syft_proto.types.torch.v1.parameter_pb2 import Parameter as ParameterPB
