+++ syft/workers/node_client.py
@@ -5,7 +5,7 @@ from urllib.parse import urlparse
 
 # Syft imports
 from syft.serde import serialize
-from syft.messaging.plan import Plan
+from syft.execution.plan import Plan
 from syft.codes import REQUEST_MSG, RESPONSE_MSG
 from syft.federated.federated_client import FederatedClient
 from syft.workers.websocket_client import WebsocketClientWorker
