+++ test/torch/pointers/test_pointer_plan.py
@@ -2,7 +2,7 @@ import torch as th
 import syft as sy
 
 from syft.generic.pointers.pointer_plan import PointerPlan
-from syft.messaging.plan import Plan
+from syft.execution.plan import Plan
 
 
 def test_create_pointer_to_plan(hook, workers):
